Rewrite modules to the new recommended Boost practice
