+++ CMakeLists.txt
@@ -4,20 +4,37 @@
 # Distributed under the Boost Software License, Version 1.0.
 # https://www.boost.org/LICENSE_1_0.txt
 
-cmake_minimum_required(VERSION 3.5...3.16)
+cmake_minimum_required(VERSION 3.5...3.31)
 
 project(boost_pfr VERSION "${BOOST_SUPERPROJECT_VERSION}" LANGUAGES CXX)
 
-add_library(boost_pfr INTERFACE)
-add_library(Boost::pfr ALIAS boost_pfr)
-
-target_include_directories(boost_pfr INTERFACE include)
-
-if (CMAKE_VERSION VERSION_GREATER_EQUAL "3.28.0" AND BUILD_MODULE)
-    add_subdirectory(module)
+if (CMAKE_VERSION VERSION_GREATER_EQUAL "3.28.0" AND BOOST_USE_MODULES)
+    add_library(boost_pfr STATIC)
+    target_compile_features(boost_pfr PUBLIC cxx_std_23)
+    target_sources(boost_pfr PUBLIC
+        FILE_SET modules_public TYPE CXX_MODULES FILES
+            ${CMAKE_CURRENT_LIST_DIR}/modules/pfr.cppm
+    )
+    set_target_properties(boost_pfr PROPERTIES CXX_MODULE_STD 1)
+    target_compile_definitions(boost_pfr PUBLIC BOOST_USE_MODULES)
+    target_include_directories(boost_pfr PUBLIC ${CMAKE_CURRENT_LIST_DIR}/include)
+else()
+    add_library(boost_pfr INTERFACE)
+    target_include_directories(boost_pfr INTERFACE ${CMAKE_CURRENT_LIST_DIR}/include)
 endif()
+
+add_library(Boost::pfr ALIAS boost_pfr)
 
 if (BUILD_TESTING AND EXISTS "${CMAKE_CURRENT_SOURCE_DIR}/test/CMakeLists.txt")
   add_subdirectory(test)
+endif()
+
+if (CMAKE_VERSION VERSION_GREATER_EQUAL "3.28.0" AND BOOST_USE_MODULES AND BUILD_TESTING)
+    add_executable(boost_pfr_module_usage modules/usage_sample.cpp)
+    target_link_libraries(boost_pfr_module_usage PRIVATE Boost::pfr)
+
+    # Make sure that mixing includes and imports is fine for different TU
+    add_executable(boost_pfr_module_usage_mu modules/usage_test_mu1.cpp modules/usage_test_mu2.cpp)
+    target_link_libraries(boost_pfr_module_usage_mu PRIVATE Boost::pfr)
 endif()
