+++ include/boost/pfr/config.hpp
@@ -8,7 +8,7 @@
 #define BOOST_PFR_CONFIG_HPP
 #pragma once
 
-#if __cplusplus >= 201402L || (defined(_MSC_VER) && defined(_MSVC_LANG) && _MSC_VER > 1900)
+#if !defined(BOOST_USE_MODULES)
 #include <type_traits> // to get non standard platform macro definitions (__GLIBCXX__ for example)
 #endif
 
@@ -70,8 +70,10 @@
 #endif
 
 #ifndef BOOST_PFR_USE_STD_MAKE_INTEGRAL_SEQUENCE
+#   if defined(BOOST_USE_MODULES)
+#       define BOOST_PFR_USE_STD_MAKE_INTEGRAL_SEQUENCE 1
 // Assume that libstdc++ since GCC-7.3 does not have linear instantiation depth in std::make_integral_sequence
-#   if defined( __GLIBCXX__) && __GLIBCXX__ >= 20180101
+#   elif defined( __GLIBCXX__) && __GLIBCXX__ >= 20180101
 #       define BOOST_PFR_USE_STD_MAKE_INTEGRAL_SEQUENCE 1
 #   elif defined(_MSC_VER)
 #       define BOOST_PFR_USE_STD_MAKE_INTEGRAL_SEQUENCE 1
@@ -144,13 +146,5 @@
 #endif
 
 #undef BOOST_PFR_NOT_SUPPORTED
-
-#ifndef BOOST_PFR_BEGIN_MODULE_EXPORT
-#   define BOOST_PFR_BEGIN_MODULE_EXPORT
-#endif
-
-#ifndef BOOST_PFR_END_MODULE_EXPORT
-#   define BOOST_PFR_END_MODULE_EXPORT
-#endif
 
 #endif // BOOST_PFR_CONFIG_HPP
