+++ include/boost/pfr/detail/size_array.hpp
@@ -9,7 +9,7 @@
 
 #include <boost/pfr/detail/config.hpp>
 
-#ifdef BOOST_PFR_HAS_STD_MODULE
+#if defined(BOOST_USE_MODULES)
 import std;
 #else
 #include <cstddef>
