+++ include/boost/pfr/detail/stdtuple.hpp
@@ -9,14 +9,14 @@
 
 #include <boost/pfr/detail/config.hpp>
 
-#ifdef BOOST_PFR_HAS_STD_MODULE
+#include <boost/pfr/detail/sequence_tuple.hpp>
+
+#if defined(BOOST_USE_MODULES)
 import std;
 #else
 #include <utility>      // metaprogramming stuff
 #include <tuple>
 #endif
-
-#include <boost/pfr/detail/sequence_tuple.hpp>
 
 namespace boost { namespace pfr { namespace detail {
