+++ include/boost/pfr/functors.hpp
@@ -9,6 +9,10 @@
 
 #include <boost/pfr/detail/config.hpp>
 
+#if defined(BOOST_USE_MODULES) && !defined(BOOST_PFR_INTERFACE_UNIT)
+import boost.pfr;
+#else
+
 #include <boost/pfr/ops.hpp>
 
 #include <boost/pfr/detail/functional.hpp>
@@ -34,8 +38,6 @@
 ///
 /// \b Synopsis:
 namespace boost { namespace pfr {
-
-BOOST_PFR_BEGIN_MODULE_EXPORT
 
 ///////////////////// Comparisons
 
@@ -218,8 +220,8 @@ template <class T> struct hash {
     }
 };
 
-BOOST_PFR_END_MODULE_EXPORT
-
 }} // namespace boost::pfr
+
+#endif  // #if defined(BOOST_USE_MODULES) && !defined(BOOST_PFR_INTERFACE_UNIT)
 
 #endif // BOOST_PFR_FUNCTORS_HPP
