+++ include/boost/pfr/traits.hpp
@@ -9,8 +9,17 @@
 
 #include <boost/pfr/detail/config.hpp>
 
+#if defined(BOOST_USE_MODULES) && !defined(BOOST_PFR_INTERFACE_UNIT)
+import boost.pfr;
+#else
+
 #include <boost/pfr/detail/possible_reflectable.hpp>
+
+#if defined(BOOST_USE_MODULES)
+import std;
+#else
 #include <type_traits>
+#endif
 
 /// \file boost/pfr/traits.hpp
 /// Contains traits \forcedlink{is_reflectable} and \forcedlink{is_implicitly_reflectable} for detecting an ability to reflect type.
@@ -18,8 +27,6 @@
 /// \b Synopsis:
 
 namespace boost { namespace pfr {
-
-BOOST_PFR_BEGIN_MODULE_EXPORT
 
 /// Has a static const member variable `value` when it is known that type T can or can't be reflected using Boost.PFR; otherwise, there is no member variable.
 /// Every user may (and in some difficult cases - should) specialize is_reflectable on his own.
@@ -56,9 +63,9 @@ using is_implicitly_reflectable = std::integral_constant< bool, boost::pfr::deta
 template<class T, class WhatFor>
 constexpr bool is_implicitly_reflectable_v = is_implicitly_reflectable<T, WhatFor>::value;
 
-BOOST_PFR_END_MODULE_EXPORT
-
 }} // namespace boost::pfr
+
+#endif  // #if defined(BOOST_USE_MODULES) && !defined(BOOST_PFR_INTERFACE_UNIT)
 
 #endif // BOOST_PFR_TRAITS_HPP
