+++ modules/CMakeLists.txt
@@ -1,0 +1,11 @@
+# Copyright (c) 2016-2025 Antony Polukhin
+#
+# Distributed under the Boost Software License, Version 1.0.
+# https://www.boost.org/LICENSE_1_0.txt
+
+cmake_minimum_required(VERSION 3.28)
+
+function (_add_boost_pfr_module NAME)
+endfunction()
+
+
