+++ modules/usage_test_mu1.cpp
@@ -6,27 +6,22 @@
 // To compile manually use a command like the folowing:
 // clang++ -std=c++20 -fmodule-file=pfr.pcm pfr.pcm usage_sample.cpp
 
-#include <iostream>
-#include <iomanip>
+#include <boost/pfr.hpp>
 
-import Boost.PFR;
+import std;
 
 struct some_person {
     std::string name;
     unsigned birth_year;
 };
 
-void mu1_act();
-
-int main() {
-    mu1_act();
-
-    some_person val{"Joseph Brodsky", 1940};
+void mu1_act() {
+    some_person val{"Edgar Allan Poe", 1809};
 
     std::cout << boost::pfr::get<0>(val)                // No macro!
         << " was born in " << boost::pfr::get<1>(val);  // Works with any aggregate!
 
-    std::cout << '\n' << boost::pfr::io(val);
+    std::cout << '\n' << boost::pfr::io(val);           // Outputs: {"Edgar Allan Poe", 1809}
     std::cout << "\n." << boost::pfr::get_name<0, some_person>()
-        << '=' << val.name << '\n';
+        << '=' << val.name << '\n';                     // Outputs: .name=Edgar Allan Poe
 }
